refactor(xod-js,xod-project): move defaultizers back into /test/helpers and import it from xod-js by full path (xod-project/test/helpers). Also add `def` to transpile functions
