+++ packages/xod-js/src/target-espruino.js
@@ -1,3 +1,4 @@
+import { def } from './types';
 import runTranspile from './transpiler';
 
 import espruinoLauncher from '../platform/espruino/launcher';
@@ -5,16 +6,17 @@ import espruinoLauncher from '../platform/espruino/launcher';
 /**
  * Runs transpilation for Espruino target.
  *
- * @function transpile
+ * @function transpileForEspruino
  * @param {Project} project Whole project (v2)
  * @param {Path} path Path of entry-point patch
  * @returns {String} Transpiled code
  */
-export default function transpile(project, path) {
-  return runTranspile({
+export default def(
+  'transpileForEspruino :: Project -> PatchPath -> String',
+  (project, path) => runTranspile({
     project,
     path,
     impls: ['espruino', 'js'],
     launcher: espruinoLauncher,
-  });
-}
+  })
+);
