+++ packages/xod-js/src/target-nodejs.js
@@ -1,3 +1,4 @@
+import { def } from './types';
 import runTranspile from './transpiler';
 
 import nodejsLauncher from '../platform/nodejs/launcher';
@@ -5,16 +6,17 @@ import nodejsLauncher from '../platform/nodejs/launcher';
 /**
  * Runs transpilation for NodeJS target.
  *
- * @function transpile
+ * @function transpileForNodeJS
  * @param {Project} project Whole project (v2)
  * @param {Path} path Path of entry-point patch
  * @returns {String} Transpiled code
  */
-export default function transpile(project, path) {
-  return runTranspile({
+export default def(
+  'transpileForNodeJS :: Project -> PatchPath -> String',
+  (project, path) => runTranspile({
     project,
     path,
     impls: ['nodejs', 'js'],
     launcher: nodejsLauncher,
-  });
-}
+  })
+);
