+++ packages/xod-js/src/types.js
@@ -1,0 +1,5 @@
+import HMDef from 'hm-def';
+import { env } from 'xod-project';
+
+export const def = HMDef.create({ checkTypes: true, env });
+export default def;
