+++ packages/xod-project/test/helpers.js
@@ -1,8 +1,6 @@
 import R from 'ramda';
 import { Either } from 'ramda-fantasy';
 
-import * as defaults from '../src/defaults';
-
 export const expectEither = R.curry((testFunction, object) => {
   Either.either(
     (err) => { throw err; },
@@ -91,12 +89,52 @@ export const expectOptionalNumberSetter = R.curry((expect, method, propName) => 
 // Defaultizers
 //-----------------------------------------------------------------------------
 
-export const defaultizeLink = defaults.defaultizeLink;
+export const defaultizeLink = R.merge({
+  id: '$$defaultLinkId',
+  input: { nodeId: '$$defaultInputNodeId', pinKey: '$$defaultInputPin' },
+  output: { nodeId: '$$defaultOutputNodeId', pinKey: '$$defaultOutputPin' },
+});
 
-export const defaultizeNode = defaults.defaultizeNode;
+export const defaultizeNode = R.merge({
+  id: '$$defaultNodeId',
+  position: { x: 0, y: 0 },
+  type: '@/defaultType',
+  label: '',
+});
 
-export const defaultizePin = defaults.defaultizePin;
+export const defaultizePin = R.merge({
+  key: '$$defaultPinKey',
+  direction: '$$defaultDirection',
+  label: '$$defaultLabel',
+  type: 'number',
+  value: 0,
+  order: 0,
+  description: '$$defaultDesription',
+});
 
-export const defaultizePatch = defaults.defaultizePatch;
+export const defaultizePatch = R.compose(
+  R.evolve({
+    nodes: R.map(defaultizeNode),
+    links: R.map(defaultizeLink),
+    impls: R.identity,
+    pins: R.map(defaultizePin),
+  }),
+  R.merge({
+    nodes: {},
+    links: {},
+    impls: {},
+    pins: {},
+  })
+);
 
-export const defaultizeProject = defaults.defaultizeProject;
+export const defaultizeProject = R.compose(
+  R.evolve({
+    patches: R.map(defaultizePatch),
+  }),
+  R.merge({
+    authors: [],
+    license: '',
+    description: '',
+    patches: {},
+  })
+);
